Refactor - rename db-query.js to db-connect.js
