+++ todos.js
@@ -12,7 +12,7 @@ const PgPersistence = require("./lib/pg-persistence.js");
 //const {persistence} = require("./lib/get-config.js");
 //const Persistence = require(persistence);
 const catchError = require("./lib/catch-error");
-const {getClient} = require("./lib/db-query");
+const {getClient} = require("./lib/db-connect");
 app.set("views", "./views");
 app.set("view engine", "pug");
